--- Sources/HealthChecks/Extensions/Application/Application+Formatting.swift
@@ -1,39 +1,0 @@
-// FS App Health Checks
-// Copyright (C) 2024  FREEDOM SPACE, LLC
-
-//
-//  This program is free software: you can redistribute it and/or modify
-//  it under the terms of the GNU Affero General Public License as published
-//  by the Free Software Foundation, either version 3 of the License, or
-//  (at your option) any later version.
-//
-//  This program is distributed in the hope that it will be useful,
-//  but WITHOUT ANY WARRANTY; without even the implied warranty of
-//  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-//  GNU Affero General Public License for more details.
-//
-//  You should have received a copy of the GNU Affero General Public License
-//  along with this program.  If not, see <https://www.gnu.org/licenses/>.
-
-//
-//  Application+Formatting.swift
-//  fs-app-health-checks
-//
-//  Created by Mykola Buhaiov on 27.03.2026.
-//
-
-import Vapor
-
-extension Application {
-    // MARK: - Formatting
-
-    /// ISO 8601-like formatter (`yyyy-MM-dd'T'HH:mm:ss.SSS`)
-    ///
-    /// - Warning: `DateFormatter` is not thread-safe.
-    /// - Performance: Creates new instance each call.
-    public var dateTimeISOFormat: DateFormatter {
-        let formatter = DateFormatter()
-        formatter.dateFormat = Constants.dateFormat
-        return formatter
-    }
-}
