--- Sources/HealthChecks/Extensions/Application/Application+Requests.swift
@@ -1,59 +1,0 @@
-// FS App Health Checks
-// Copyright (C) 2024  FREEDOM SPACE, LLC
-
-//
-//  This program is free software: you can redistribute it and/or modify
-//  it under the terms of the GNU Affero General Public License as published
-//  by the Free Software Foundation, either version 3 of the License, or
-//  (at your option) any later version.
-//
-//  This program is distributed in the hope that it will be useful,
-//  but WITHOUT ANY WARRANTY; without even the implied warranty of
-//  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-//  GNU Affero General Public License for more details.
-//
-//  You should have received a copy of the GNU Affero General Public License
-//  along with this program.  If not, see <https://www.gnu.org/licenses/>.
-
-//
-//  Application+Requests.swift
-//  fs-app-health-checks
-//
-//  Created by Mykola Buhaiov on 27.03.2026.
-//
-
-import Vapor
-
-extension Application {
-    // MARK: - Request Context
-
-    /// Storage key for PostgreSQL request context.
-    public struct PsqlRequestKey: StorageKey {
-        public typealias Value = PsqlRequestSendable
-    }
-    /// PostgreSQL request context for the current request.
-    public var psqlRequest: PsqlRequestSendable? {
-        get { storage[PsqlRequestKey.self] }
-        set { storage[PsqlRequestKey.self] = newValue }
-    }
-
-    /// Storage key for Redis request context.
-    public struct RedisRequestKey: StorageKey {
-        public typealias Value = RedisRequestSendable
-    }
-    /// Redis request context for the current request.
-    public var redisRequest: RedisRequestSendable? {
-        get { storage[RedisRequestKey.self] }
-        set { storage[RedisRequestKey.self] = newValue }
-    }
-
-    /// Storage key for MongoDB request context.
-    public struct MongoRequestSendableKey: StorageKey {
-        public typealias Value = MongoRequestSendable
-    }
-    /// MongoDB request context for the current request.
-    public var mongoRequest: MongoRequestSendable? {
-        get { storage[MongoRequestSendableKey.self] }
-        set { storage[MongoRequestSendableKey.self] = newValue }
-    }
-}
